Create i-engage HTML & JS
Encourage your child to listen to the story carefully and observe the scenes. Instruct the child to identify words that are highlighted and match them to the relevant pictures given in the side panel. Each story enhances the child’s social skills by inspiring them to observe their surrounding with more attention to detail and make deeper connections with not just human beings but every other object in the environment. Happy reading!